Updates 11/02

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: ghenriqu <ghenriqu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/10 00:25:43 by ghenriqu          #+#    #+#             */
-/*   Updated: 2021/02/10 01:01:48 by ghenriqu         ###   ########.fr       */
+/*   Updated: 2021/02/11 03:06:28 by ghenriqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *str, int c, size_t n)
 {
-	unsigned int count;
+	size_t count;
 
 	count = 0;
 	while (count < n)
