+++ libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: ghenriqu <ghenriqu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/10 00:25:50 by ghenriqu          #+#    #+#             */
-/*   Updated: 2021/02/10 04:59:26 by ghenriqu         ###   ########.fr       */
+/*   Updated: 2021/02/11 02:15:23 by ghenriqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int     ft_memcmp(const void *str1, const void *str2, size_t n)
 {
-	unsigned int count;
+	size_t count;
 	int value;
 
 	count = 0;
