+++ libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: ghenriqu <ghenriqu@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/10 00:25:59 by ghenriqu          #+#    #+#             */
-/*   Updated: 2021/02/10 00:26:00 by ghenriqu         ###   ########.fr       */
+/*   Updated: 2021/02/11 02:15:19 by ghenriqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
-    unsigned int count;
+	size_t count;
 
-    count = 0;
-    while (count < n)
-    {
-        ((unsigned char *)dest)[count] = ((unsigned char *)src)[count];
-        count++;
-    }
-    return (dest);
+	count = 0;
+	while (count < n)
+	{
+		((unsigned char *)dest)[count] = ((unsigned char *)src)[count];
+		count++;
+	}
+	return (dest);
 }
